v0.0.4 Renamed file folders and respective references
